Fix response_time metric type

+++ worker.go
@@ -115,7 +115,7 @@ func (Worker) handleMessage(msg *sqs.Message) error {
 		defer func() {
 			// Get difference in milliseconds
 			diff := (time.Now().UnixNano() - now.UnixNano()) / 1000000
-			go StatsClient.Histogram("response_time", diff, nil)
+			go StatsClient.Histogram("response_time", float64(diff), nil)
 		}()
 	}
 	if res.StatusCode > 299 || res.StatusCode < 200 {
